move base prop class into another base file
