+++ chebai_graph/preprocessing/properties/__init__.py
@@ -2,10 +2,10 @@
 # but it has to be imported after properties module, to avoid circular imports
 # This is because augmented properties module imports from properties module
 # isort: off
+
+from .base import MolecularProperty, AtomProperty, BondProperty
+
 from .properties import (
-    MolecularProperty,
-    AtomProperty,
-    BondProperty,
     AtomType,
     NumAtomBonds,
     AtomCharge,
